converter: defect D -- an above-uint32 literal reaches a NATIVE-WIDTH unsigned type
`var word uintptr = 0x0102030405060708` emitted `word = 0x0102030405060708UL`.
The `UL` suffix makes it a C# ulong, and golib's uintptr converts from ulong only
EXPLICITLY -- its implicit sources are nuint, uint8/16/32, char, UntypedInt and
NilType -- so the initializer is CS0266. Found by i9 minting the golden on
d839cb1d7, on a guard row this stack itself added; ruled to me in 944cce634.

IT IS A DRIFT BETWEEN TWO BRANCHES OF ONE RULE, not a missing rule.

The literal path has two above-MaxUint32 branches, split by how the value parses:

  signed parse    value <= MaxInt64   emitted a BARE `UL`
  unsigned parse  value >  MaxInt64   emitted `(nuint)` + `UL`

The SECOND branch already states the rule exactly -- a uint64 context takes the
plain literal because the [GoType] wrappers over uint64 convert implicitly from
ulong, while a native-width unsigned context (Go uint/uintptr -> C# nuint) keeps
the `(nuint)` cast because a bare ulong has no implicit conversion to nuint. The
FIRST branch never got it, so the defect lives exactly in the range owned by the
branch without the rule, which is why nothing had met it before: a Go constant
above MaxUint32 in a uintptr context, at or below MaxInt64.

So the fix is ONE predicate that both branches call, rather than a second copy of
the rule in the branch that lacked it. The comment moves to the helper with it.
Two derivations of one predicate is how this happened; there is now one.

  before   word = 0x0102030405060708UL;          CS0266
  after    word = (nuint)0x0102030405060708UL;   uintptr's implicit nuint binds

The `(nuint)` cast is a non-constant unchecked conversion, which compiles where
the constant conversion does not -- the mechanism the sibling branch documents.

GATES (base = d839cb1d7):
  converter go test ./... -count=1   ok go2cs 184.477s, exit 0
  RED CONTROL                        neutering the helper returns the bare `UL`
                                     to the guard's own emission; source restore
                                     byte-identical
  guard row                          IS the red control, per 944cce634 -- the row
                                     that unmasked this is already in the project

FOOTPRINT PREDICTED ZERO, and derived rather than guessed: the corpus compiles
clean at 307/307, so no corpus site can be taking the signed-parse branch in a
native-width unsigned context -- such a site would be emitting a bare ulong into
a uint/uintptr and would be CS0266 today. The only contexts that reach that
branch in the corpus are uint64 ones, where the helper returns "" and the
emission is byte-identical. A differing corpus file would be a finding about the
compile gate rather than about this fix.

NOT RUN: Compile and Output phases and the golden -- .NET, and this box has no
.NET 10 SDK. i9 re-mints on this tip.

+++ src/tests/Behavioral/SwitchPointerSentinelCase/main.cs
@@ -78,7 +78,7 @@ internal static void Main() {
     ref var elsewhere = ref heap(new mu(), out var Ꮡelsewhere);
     fmt.Println(preferLowLatency(ᏑtheSched.of(schedt.Ꮡlock)), preferLowLatency(Ꮡelsewhere), preferLowLatency(nil));
     ref var word = ref heap(new uintptr(), out var Ꮡword);
-    word = 0x0102030405060708UL;
+    word = (nuint)0x0102030405060708UL;
     fmt.Println(key8(Ꮡword).Value, key8Last(Ꮡword).Value);
 }
